Rest of 9/16 fixes

+++ Source/WebKit/UIProcess/Cocoa/SystemPreviewControllerCocoa.mm
@@ -179,7 +179,7 @@ SOFT_LINK_CLASS(QuickLook, QLItem);
     UIViewController *presentingViewController = [self presentingViewController];
     if (presentingViewController) {
         if (_linkRect.isEmpty())
-            *contentRect = {CGPointZero, {presentingViewController.view.frame.size.width / 2.0, presentingViewController.view.frame.size.height / 2.0}};
+            *contentRect = {CGPointZero, {static_cast<CGFloat>(presentingViewController.view.frame.size.width / 2.0), static_cast<CGFloat>(presentingViewController.view.frame.size.height / 2.0)}};
         else {
             WebCore::IntRect screenRect = _previewController->page().syncRootViewToScreen(_linkRect);
             *contentRect = { CGPointZero, { static_cast<CGFloat>(screenRect.width()), static_cast<CGFloat>(screenRect.height()) } };
